Changed keybinds from maps to functions and no longer streaming.
Using functions instead of maps allows use to take advantage of pattern
matching when describing a modes functions e.g. we can now pattern match
for (KChar c) which was not possible before. Will be harder to extend
although not impossible.

Removed the streaming as this was needlessly complicated when trying to
load new buffers.