Created entirely new buffer added tabs. Works somewhat.
A radical structure change involving implementing a new
easier to work with (and more efficient) text buffer.
Also an interpreter for the TextT free monad has been
implemented allowing the use of tabs. The basic structure
works well although in the future I will be updating
the interpreter to work with split windows and 'views'.