Refactored the vertical movements to use less buffer queries.
I realised we can use a maybe value to store the preferred column.
When a line move is requested we first check whether preferredCol
is Nothing; if so, we must query the current col and set it as
the preferredCol; and, when preferredCol is set, we can use that
value for line movement. This requires that any movement that
changes the column must reset preferredCol; although, this
is of little cost.